完善我的世界 模型类型接口

+++ tool-common/src/main/java/com/gxlirong/tool/common/api/CommonResult.java
@@ -9,14 +9,14 @@ import lombok.Data;
  */
 @Data
 public class CommonResult<T> {
-    private long code;
+    private int code;
     private String message;
     private T data;
 
     protected CommonResult() {
     }
 
-    public CommonResult(long code, String message, T data) {
+    public CommonResult(int code, String message, T data) {
         this.code = code;
         this.message = message;
         this.data = data;
